chore(route) Add test for routes - Add test for route component - Rename view to pages [Delivers #163897632]
