+++ src/App.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Routes from './routes/routes';
+import Routes from './routes/Routes';
 
 const App = () => {
   return (
